ingestion_functions/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-storage-blob
pypdf
python-docx
openai
azure-search-documents
python-dotenv
python-jose requests